+++ sanpy/bDetection.py
@@ -36,7 +36,7 @@ import glob
 import copy
 from collections import OrderedDict
 
-from matplotlib.font_manager import json_load
+# from matplotlib.font_manager import json_load
 
 # from colin.stochAnalysis import load